chore: Apply LLVM coding standards cleanup across codebase (#204)
## Summary

- Prefer pre-increment (`++i`) over post-increment (`i++`) per [LLVM
coding
standards](https://llvm.org/docs/CodingStandards.html#prefer-preincrement)
- Use `size_t` instead of `int` for loop variables compared against
`size_t` values (`pointwise_node.h`, `test_memstream.cpp`)
- Replace doxygen-style `///` comments with `//` comments to match
project conventions (`dllib.h`, `reduction_node.h`)

29 files changed, 88 insertions(+), 88 deletions(-) — pure mechanical
substitutions with no behavioral changes.


---------

+++ tests/test_memstream.cpp
@@ -110,8 +110,8 @@ TEST_CASE("MemStream large writes", "[memstream]") {
     MemStream ms;
     REQUIRE(ms.isValid());
 
-    for (int i = 0; i < 1000; ++i) {
-      fprintf(ms, "%d", i % 10);
+    for (size_t i = 0; i < 1000; ++i) {
+      fprintf(ms, "%zu", i % 10);
     }
     std::string result = *ms.str();
     REQUIRE(result.size() == 1000);
